# customer-segmentation-api/docs/swagger.yaml
basePath: /
definitions:
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  models.Insight:
    properties:
      action:
        example: Email campaign
        type: string
      priority:
        example: high
        type: string
    type: object
  models.LoginRequest:
    properties:
      password:
        example: test-password
        type: string
      username:
        example: test-user
        type: string
    type: object
  models.RegisterRequest:
    properties:
      email:
        type: string
      full_name:
        maxLength: 100
        minLength: 2
        type: string
      password:
        maxLength: 100
        minLength: 6
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - full_name
    - password
    - username
    type: object
  models.SegmentRequest:
    type: object
  models.SegmentResponse:
    properties:
      confidence:
        example: 0.85
        type: number
      customer_id:
        example: "12345"
        type: string
      insights:
        items:
          $ref: '#/definitions/models.Insight'
        type: array
      recommended_offer:
        example: 15% discount
        type: string
      segment:
        example: tech-savvy
        type: string
    type: object
  models.UserProfile:
    properties:
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      role:
        type: string
      username:
        type: string
    type: object
host: localhost:8008
info:
  contact: {}
  description: MiniBatchKMeans ve Ollama LLM ile müşteri segmentasyonu yapan Go API
  title: Customer Segmentation API
  version: "1.0"
paths:
  /api/login:
    post:
      consumes:
      - application/json
      description: Login with username and password to get JWT token
      parameters:
      - description: User credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: User login
      tags:
      - auth
  /api/profile:
    get:
      description: Get current user profile information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserProfile'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - user
  /api/register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: User registration data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.UserProfile'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
  /api/segment:
    post:
      consumes:
      - application/json
      description: Müşteri verilerini segmentler ve Python servisine yönlendirir
      parameters:
      - description: Müşteri segmentasyon isteği
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SegmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SegmentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Müşteri segmentasyonu
      tags:
      - segment
securityDefinitions:
  BearerAuth:
    description: 'Enter your Bearer token in the format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"